Bug 2035813 - [devtools] Don't call removeExtensionSidebar in ExtensionSidebar#destroy when called from Inspector#destroy. r=devtools-reviewers,jdescottes.
Differential Revision: https://phabricator.services.mozilla.com/D297333

UltraBlame original commit: fd4b90f4afb0a57eba5bfdae23d9002ee13995a9

+++ devtools/client/inspector/extensions/extension-sidebar.js
@@ -129,15 +129,20 @@ class ExtensionSidebar extends EventEmitter {
 
 
 
-  destroy() {
+
+
+
+  destroy({ fromInspectorDestroy } = {}) {
     if (this.destroyed) {
       throw new Error(
         `ExtensionSidebar instances cannot be destroyed more than once`
       );
     }
 
-    
-    this.store.dispatch(removeExtensionSidebar(this.id));
+    if (fromInspectorDestroy !== true) {
+      
+      this.store.dispatch(removeExtensionSidebar(this.id));
+    }
 
     this.inspector = null;
     this.store = null;
